+++ src/hooks/useDeleteVideo.ts
@@ -1,4 +1,5 @@
 import { useApolloClient } from '@apollo/client'
+import { useCallback } from 'react'
 
 import { useConfirmationModal } from '@/providers/confirmationModal'
 import { useJoystream } from '@/providers/joystream'
@@ -16,47 +17,53 @@ export const useDeleteVideo = () => {
 
   const client = useApolloClient()
 
-  const deleteVideo = (videoId: string, onDeleteVideo?: () => void) => {
-    openDeleteVideoDialog({
-      title: 'Delete this video?',
-      description:
-        'You will not be able to undo this. Deletion requires a blockchain transaction to complete. Currently there is no way to remove uploaded video assets.',
-      primaryButton: {
-        text: 'Delete video',
-        variant: 'destructive',
-        onClick: () => {
-          confirmDeleteVideo(videoId, () => onDeleteVideo?.())
-          closeDeleteVideoDialog()
-        },
-      },
-      secondaryButton: {
-        text: 'Cancel',
-        onClick: () => {
-          closeDeleteVideoDialog()
-        },
-      },
-      iconType: 'warning',
-    })
-  }
+  const confirmDeleteVideo = useCallback(
+    async (videoId: string, onTxSync?: () => void) => {
+      if (!joystream) {
+        return
+      }
 
-  const confirmDeleteVideo = async (videoId: string, onTxSync?: () => void) => {
-    if (!joystream) {
-      return
-    }
+      handleTransaction({
+        txFactory: (updateStatus) => joystream.deleteVideo(videoId, activeMemberId, updateStatus),
+        onTxSync: async () => {
+          removeVideoFromCache(videoId, client)
+          removeAssetsWithParentFromUploads('video', videoId)
+          onTxSync?.()
+        },
+        successMessage: {
+          title: 'Video successfully deleted!',
+          description: 'Your video was marked as deleted and it will no longer show up on Joystream.',
+        },
+      })
+    },
+    [activeMemberId, client, handleTransaction, joystream, removeAssetsWithParentFromUploads]
+  )
 
-    handleTransaction({
-      txFactory: (updateStatus) => joystream.deleteVideo(videoId, activeMemberId, updateStatus),
-      onTxSync: async () => {
-        removeVideoFromCache(videoId, client)
-        removeAssetsWithParentFromUploads('video', videoId)
-        onTxSync?.()
-      },
-      successMessage: {
-        title: 'Video successfully deleted!',
-        description: 'Your video was marked as deleted and it will no longer show up on Joystream.',
-      },
-    })
-  }
+  const deleteVideo = useCallback(
+    (videoId: string, onDeleteVideo?: () => void) => {
+      openDeleteVideoDialog({
+        title: 'Delete this video?',
+        description:
+          'You will not be able to undo this. Deletion requires a blockchain transaction to complete. Currently there is no way to remove uploaded video assets.',
+        primaryButton: {
+          text: 'Delete video',
+          variant: 'destructive',
+          onClick: () => {
+            confirmDeleteVideo(videoId, () => onDeleteVideo?.())
+            closeDeleteVideoDialog()
+          },
+        },
+        secondaryButton: {
+          text: 'Cancel',
+          onClick: () => {
+            closeDeleteVideoDialog()
+          },
+        },
+        iconType: 'warning',
+      })
+    },
+    [closeDeleteVideoDialog, confirmDeleteVideo, openDeleteVideoDialog]
+  )
 
   return deleteVideo
 }
